refactor: 컨텐츠 컴포넌트화
